+++ server/src/main/java/com/watchtogether/server/party/controller/PartyController.java
@@ -2,7 +2,7 @@ package com.watchtogether.server.party.controller;
 
 import com.watchtogether.server.party.domain.model.AcceptPartyForm;
 import com.watchtogether.server.party.domain.model.CreatePartyForm;
-import com.watchtogether.server.party.service.PartyService;
+import com.watchtogether.server.party.service.impl.PartyServiceImpl;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
@@ -13,17 +13,17 @@ import org.springframework.web.bind.annotation.*;
 @Tag(name = "파티(PartyController)", description = "Party API")
 @RequestMapping("/api/v1/parties")
 public class PartyController {
-    private final PartyService partyService;
+    private final PartyServiceImpl partyServiceimpl;
 
 
     @PostMapping("/create")
     public ResponseEntity<?> createParty(@RequestBody CreatePartyForm form){
-        return ResponseEntity.ok(partyService.createParty(form));
+        return ResponseEntity.ok(partyServiceimpl.createParty(form));
     }
 
     @PutMapping("/accept")
     public ResponseEntity<?> acceptParty(@RequestBody AcceptPartyForm form){
-       return ResponseEntity.ok(partyService.acceptParty(form));
+       return ResponseEntity.ok(partyServiceimpl.acceptParty(form));
     }
